+++ fractals/src/main/java/me/catmousedog/fractals/main/Settings.java
@@ -210,7 +210,7 @@ public class Settings {
 		logger.log(Level.FINER, "Settings.initFractals");
 
 		allFractals = new Fractal[] { new Mandelbrot(), new JuliaSet(), new BurningShip(), new JuliaShip(),
-				new InverseMandelbrot(), new InverseJulia(), new Multibrot() };
+				new InverseMandelbrot(), new InverseJulia(), new Multibrot(), new TestFractal() };
 
 		// scan and copy resources inside 'conrete_fractals' resource