+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/filters/PeriodicFilter.java
@@ -1,12 +1,14 @@
 package me.catmousedog.fractals.paneloperators.filters;
 
-import java.awt.Color;
+import java.awt.Font;
 
 import me.catmousedog.fractals.paneloperators.fractals.Fractal;
 import me.catmousedog.fractals.ui.components.Item;
 import me.catmousedog.fractals.ui.components.concrete.Button;
+import me.catmousedog.fractals.ui.components.concrete.Label;
 import me.catmousedog.fractals.ui.components.concrete.Padding;
 import me.catmousedog.fractals.ui.components.concrete.SliderDouble;
+import me.catmousedog.fractals.ui.components.concrete.SubTitle;
 import me.catmousedog.fractals.ui.components.concrete.TextFieldDouble;
 
 /**
@@ -19,15 +21,9 @@ public class PeriodicFilter extends Filter {
 
 	private boolean inverted;
 
-	private double r;
-	private double g;
-	private double b;
-
-	private double rf;
-	private double gf;
-	private double bf;
-
 	private Button invertjb;
+
+	private double r, g, b;
 
 	private TextFieldDouble rjtf;
 	private SliderDouble rjs;
@@ -38,6 +34,8 @@ public class PeriodicFilter extends Filter {
 	private TextFieldDouble bjtf;
 	private SliderDouble bjs;
 
+	private double rf, gf, bf;
+
 	private TextFieldDouble rfjtf;
 	private SliderDouble rfjs;
 
@@ -47,12 +45,25 @@ public class PeriodicFilter extends Filter {
 	private TextFieldDouble bfjtf;
 	private SliderDouble bfjs;
 
+	private double ro, bo, go;
+
+	private TextFieldDouble rojtf;
+	private SliderDouble rojs;
+
+	private TextFieldDouble gojtf;
+	private SliderDouble gojs;
+
+	private TextFieldDouble bojtf;
+	private SliderDouble bojs;
+
 	public PeriodicFilter(Fractal fractal) {
 		super(fractal);
 
 		Padding p5 = new Padding(5);
-		invertjb = new Button.Builder("Invert").setTip("Inverts the current colours.").setAction(a -> actionInvert())
-				.build();
+		invertjb = new Button.Builder("Invert").setTip("<html>Inverts the current colours</html>")
+				.setAction(a -> actionInvert()).build();
+
+		SubTitle amp = new SubTitle("amplitude");
 		String rTip = "<html>The amplitude for the red curve</html>";
 		rjtf = new TextFieldDouble.Builder().setLabel("red factor").setTip(rTip).build();
 		rjs = new SliderDouble.Builder().setTip(rTip).setChange(c -> changeR()).build();
@@ -62,25 +73,42 @@ public class PeriodicFilter extends Filter {
 		String bTip = "<html>The amplitude for the blue curve</html>";
 		bjtf = new TextFieldDouble.Builder().setLabel("blue factor").setTip(bTip).build();
 		bjs = new SliderDouble.Builder().setTip(bTip).setChange(c -> changeB()).build();
+
+		SubTitle fre = new SubTitle("frequency");
 		String rfTip = "<html>The frequency factor for the red component</html>";
 		rfjtf = new TextFieldDouble.Builder().setLabel("red frequency").setTip(rfTip).build();
-		rfjs = new SliderDouble.Builder().setTip(rfTip).setChange(c -> changeRf()).setMin(1).setMax(10).build();
+		rfjs = new SliderDouble.Builder().setTip(rfTip).setChange(c -> changeRf()).setMin(0.5).setMax(10.5).build();
 		String gfTip = "<html>The frequency factor for the green component</html>";
 		gfjtf = new TextFieldDouble.Builder().setLabel("green frequency").setTip(gfTip).build();
-		gfjs = new SliderDouble.Builder().setTip(gfTip).setChange(c -> changeGf()).setMin(1).setMax(10).build();
+		gfjs = new SliderDouble.Builder().setTip(gfTip).setChange(c -> changeGf()).setMin(0.5).setMax(10.5).build();
 		String bfTip = "<html>The frequency factor for the blue component</html>";
 		bfjtf = new TextFieldDouble.Builder().setLabel("blue frequency").setTip(bfTip).build();
-		bfjs = new SliderDouble.Builder().setTip(bfTip).setChange(c -> changeBf()).setMin(1).setMax(10).build();
-		items = new Item[] { invertjb, p5, rjtf, rjs, p5, gjtf, gjs, p5, bjtf, bjs, p5, rfjtf, rfjs, p5, gfjtf, gfjs,
-				p5, bfjtf, bfjs };
+		bfjs = new SliderDouble.Builder().setTip(bfTip).setChange(c -> changeBf()).setMin(0.5).setMax(10.5).build();
+
+		SubTitle off = new SubTitle("offset");
+		String roTip = "<html>The offset for the red component</html>";
+		rojtf = new TextFieldDouble.Builder().setLabel("red offset").setTip(roTip).build();
+		rojs = new SliderDouble.Builder().setTip(roTip).setChange(c -> changeRo()).setMin(0).setMax(1).build();
+		String goTip = "<html>The offset for the green component</html>";
+		gojtf = new TextFieldDouble.Builder().setLabel("green offset").setTip(goTip).build();
+		gojs = new SliderDouble.Builder().setTip(goTip).setChange(c -> changeGo()).setMin(0).setMax(1).build();
+		String boTip = "<html>The offset for the blue component</html>";
+		bojtf = new TextFieldDouble.Builder().setLabel("blue offset").setTip(boTip).build();
+		bojs = new SliderDouble.Builder().setTip(boTip).setChange(c -> changeBo()).setMin(0).setMax(1).build();
+
+		items = new Item[] { invertjb, p5, amp, rjtf, rjs, gjtf, gjs, bjtf, bjs, p5, fre, rfjtf, rfjs, gfjtf, gfjs,
+				bfjtf, bfjs, p5, off, rojtf, rojs, gojtf, gojs, bojtf, bojs };
 
 		inverted = false;
 		r = 1;
 		b = 1;
 		g = 1;
-		rf = 1;
-		gf = 1;
-		bf = 1;
+		rf = 0.5;
+		gf = 0.5;
+		bf = 0.5;
+		ro = 0;
+		bo = 0;
+		go = 0;
 	}
 
 	/**
@@ -98,6 +126,9 @@ public class PeriodicFilter extends Filter {
 		rf = filter.rf;
 		gf = filter.gf;
 		bf = filter.bf;
+		ro = filter.ro;
+		go = filter.go;
+		bo = filter.bo;
 	}
 
 	@Override
@@ -105,11 +136,11 @@ public class PeriodicFilter extends Filter {
 		double v = V.doubleValue();
 		if (inverted)
 			v = 1.0 - v;
-		return new Color(curve(r, rf, v), curve(g, gf, v), curve(b, bf, v)).getRGB();
+		return 0xff000000 | curve(r, rf, ro, v) << 16 | curve(g, gf, go, v) << 8 | curve(b, bf, bo, v) << 0;
 	}
 
-	private int curve(double a, double f, double v) {
-		return (int) (a * 127.5 * (1 - Math.cos(f * v * Math.PI)));
+	private int curve(double a, double f, double o, double v) {
+		return (int) (a * 127.5 * (1 - Math.cos((f * v - o) * 2 * Math.PI)));
 	}
 
 	@Override
@@ -120,6 +151,9 @@ public class PeriodicFilter extends Filter {
 		rf = rfjtf.saveAndGet();
 		gf = gfjtf.saveAndGet();
 		bf = bfjtf.saveAndGet();
+		ro = rojtf.saveAndGet();
+		go = gojtf.saveAndGet();
+		bo = bojtf.saveAndGet();
 	}
 
 	@Override
@@ -136,6 +170,12 @@ public class PeriodicFilter extends Filter {
 		rfjs.setDataSafe(rf);
 		gfjs.setDataSafe(gf);
 		bfjs.setDataSafe(bf);
+		rojtf.setData(ro);
+		gojtf.setData(go);
+		bojtf.setData(bo);
+		rojs.setDataSafe(ro);
+		gojs.setDataSafe(go);
+		bojs.setDataSafe(bo);
 	}
 
 	@Override
@@ -148,6 +188,7 @@ public class PeriodicFilter extends Filter {
 		fractal.saveAndColour();
 	}
 
+	// amplitude
 	private void changeR() {
 		rjtf.setData(rjs.saveAndGet());
 		fractal.saveAndColour();
@@ -163,6 +204,7 @@ public class PeriodicFilter extends Filter {
 		fractal.saveAndColour();
 	}
 
+	// frequency
 	private void changeRf() {
 		rfjtf.setData(rfjs.saveAndGet());
 		fractal.saveAndColour();
@@ -178,6 +220,22 @@ public class PeriodicFilter extends Filter {
 		fractal.saveAndColour();
 	}
 
+	// offset
+	private void changeRo() {
+		rojtf.setData(rojs.saveAndGet());
+		fractal.saveAndColour();
+	}
+
+	private void changeGo() {
+		gojtf.setData(gojs.saveAndGet());
+		fractal.saveAndColour();
+	}
+
+	private void changeBo() {
+		bojtf.setData(bojs.saveAndGet());
+		fractal.saveAndColour();
+	}
+
 	@Override
 	public String informalName() {
 		return "Periodic";
@@ -190,10 +248,6 @@ public class PeriodicFilter extends Filter {
 
 	@Override
 	public String getTip() {
-		return "<html>A filter designed to be used with iterative fractals."
-				+ "<br>This filter uses a periodic function for the red, green and blue components.<br>"
-				+ "Each 'curve' has its own amplitude and frequency which can be changed.<br>"
-				+ "This filter doesn't work well with high frequencies as the aliasing effect<br>"
-				+ "of iterative fractals become more nuanced.</html>";
+		return "TODO";
 	}
 }