+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/functions/EscapeAngleFunction.java
@@ -3,6 +3,7 @@ package me.catmousedog.fractals.paneloperators.functions;
 import org.jetbrains.annotations.NotNull;
 
 import me.catmousedog.fractals.data.FractalValue;
+import me.catmousedog.fractals.paneloperators.filters.BrightnessFilter;
 import me.catmousedog.fractals.paneloperators.filters.Filter;
 import me.catmousedog.fractals.paneloperators.filters.HueFilter;
 import me.catmousedog.fractals.paneloperators.filters.LinearFilter;
@@ -14,7 +15,8 @@ public class EscapeAngleFunction extends Function {
 	public EscapeAngleFunction(Fractal fractal) {
 		super(fractal);
 
-		filters = new Filter[] { new LinearFilter(fractal), new PeriodicFilter(fractal), new HueFilter(fractal) };
+		filters = new Filter[] { new LinearFilter(fractal), new BrightnessFilter(fractal), new PeriodicFilter(fractal),
+				new HueFilter(fractal) };
 		filter = filters[0];
 	}